+++ src/main-start.js
@@ -58,7 +58,7 @@ export default (settings = {}) => {
         items.forEach(item => {
             if (item && item.exists) {
               const version = item.version ? ` (v${ item.version })` : "";
-              log.info(` - '${ item.id }'${ version } routing '${ item.route }' => port:${ item.port }`);
+              log.info(` - '${ item.id }'${ version } routing '${ item.routes.toString() }' => port:${ item.port }`);
             }
         });
         log.info("");