+++ src/main.js
@@ -2,6 +2,7 @@ import R from "ramda";
 import Promise from "bluebird";
 import uuid from "uuid";
 import shell from "shelljs";
+import Route from "./route";
 import app from "./app";
 import gateway from "./gateway";
 import log from "./log";
@@ -52,6 +53,19 @@ export default (settings = {}) => {
     userAgent,
     targetFolder: settings.targetFolder || DEFAULT_TARGET_FOLDER,
 
+
+    /**
+     * Looks up the app with the given route.
+     * @param {String} domain: The domain of the route to look up.
+     * @param {String} path:   The URL path of the route.
+     * @return {Boolean}.
+     */
+    findAppFromRoute(domain, path) {
+      const isMatchingApp = app => R.find(route => route.match(domain, path), app.routes);
+      return R.find(isMatchingApp, this.apps);
+    },
+
+
     /**
      * Adds a new application to run.
      * @param {string} id:    The unique name of the app (ID).
@@ -69,9 +83,15 @@ export default (settings = {}) => {
       if (R.find(item => item.id === id, this.apps)) {
         throw new Error(`An app with the ID '${ id }' has already been registered.`);
       }
-      if (R.find(item => item.route.toString() === route, this.apps)) {
-        throw new Error(`An app with the route '${ route }' has already been registered.`);
-      }
+
+      // Ensure the route(s) are not already being used.
+      if (!R.is(Array, route)) { route = [route]; }
+      route.forEach(value => {
+            const { domain, path } = Route.parse(value)
+            if (this.findAppFromRoute(domain, path)) {
+              throw new Error(`Cannot add '${ id }' because the route '${ value }' has already been registered.`);
+            }
+          });
 
       // Create the App object.
       const port = DEFAULT_APP_PORT + (this.apps.length);