+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - API: `/api/apps` listing of all the apps (subset of the root status data-set.  
   Makes for a logically consistent API).
 #### Changed
+- Manifest: Now optionally supporting an array of routes per app.
 #### Deprecated
 #### Removed
 - Removed auto-restart when `manifest.yml` file changes.
@@ -15,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     - Now when the manifest changes force a restart using the REST api: `/api/restart`
 #### Fixed
 #### Security
+
+
 
 
 ## [1.3.0] - 2015-12-22
@@ -33,11 +36,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Updates on Github callback when `manifest.yml` changes.
 
 
+
+
 ## [1.2.0] - 2015-12-16
 #### Added
 - Added secret `tokens` for locking down the API (specified within `manifest.yml`).
 - API Action: Update
 - API Action: Restart
+
+
 
 
 ## [1.1.0] - 2015-12-14
@@ -48,6 +55,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Removed the use of the `file-system-cache` as a means of communicating when other instances are downloading an app.  Now using RabbitMQ and pub/sub for inter-container communication.
 
 
+
+
 ## [1.0.6] - 2015-11-30
 #### Added
 - Reading `targetFolder` from manifest.
@@ -56,12 +65,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - The `TARGET_FOLDER` environment variable.
 
 
+
+
 ## [1.0.1] - 2015-11-30
 #### Added
 - Version of the `app-sync` module on API status.
 #### Changed
 - API as an object on manifest (eg. `api/route`).
   This allows for more details (like `tokens`) to be associated with the API within the manifest.
+
+
 
 
 ## [1.0.0] - 2015-11-29
@@ -73,6 +86,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Changed
 - Only running NPM install when absolutely required (faster).
+
+
 
 
 ## [0.0.2] - 2015-11-8